Shannonigans/requirements.txt
numpy
matplotlib
ruptures
numba
hilbertcurve
umap-learn
scikit-learn
pandas
seaborn
tabulate
plotly
